chore: move reset logging to conftest
